[fix]解决ip为负数时会导致负载均衡计算结果下标越界的问题

+++ adapter/micro_thread/tgg/tgg_comm/tgg_bw_cache.cpp
@@ -1010,7 +1010,7 @@ void tgg_getall_bwfdx(std::vector<int64_t>& vec_bwfdx)
     }
 }
 
-int tgg_get_load_balance(std::vector<int64_t>& vec_bwfdx, int64_t ipport)
+int tgg_get_load_balance(std::vector<int64_t>& vec_bwfdx, unsigned int ipport)
 {
     // uint64_t now = get_system_ms();
     size_t count = vec_bwfdx.size();
